requirements.txt
supabase==2.4.2
requests==2.31.0
discord.py==2.3.2
groq==0.11.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pytest==8.1.1
pytest-mock==3.14.0
